+++ src/main/java/com/meoguri/linkocean/domain/bookmark/entity/BookmarkTag.java
@@ -5,7 +5,7 @@ import static lombok.AccessLevel.*;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 
-import com.meoguri.linkocean.domain.common.BaseIdEntity;
+import com.meoguri.linkocean.domain.BaseIdEntity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
