+++ src/main/java/com/meoguri/linkocean/domain/bookmark/entity/Tag.java
@@ -1,12 +1,12 @@
 package com.meoguri.linkocean.domain.bookmark.entity;
 
-import static com.meoguri.linkocean.domain.common.Preconditions.*;
+import static com.meoguri.linkocean.exception.Preconditions.*;
 import static lombok.AccessLevel.*;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 
-import com.meoguri.linkocean.domain.common.BaseIdEntity;
+import com.meoguri.linkocean.domain.BaseIdEntity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
