+++ src/main/java/com/meoguri/linkocean/domain/bookmark/entity/vo/OpenType.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.bookmark.entity;
+package com.meoguri.linkocean.domain.bookmark.entity.vo;
 
 /**
  * 북마크의 공개 범위
