+++ src/main/java/com/meoguri/linkocean/domain/category/entity/FavoriteCategory.java
@@ -6,7 +6,7 @@ import static lombok.AccessLevel.*;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 
-import com.meoguri.linkocean.domain.common.BaseIdEntity;
+import com.meoguri.linkocean.domain.BaseIdEntity;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 
 import lombok.Getter;
