+++ src/test/java/com/meoguri/linkocean/domain/bookmark/entity/ReactionTest.java
@@ -1,12 +1,10 @@
-package com.meoguri.linkocean.domain.reaction.entity;
+package com.meoguri.linkocean.domain.bookmark.entity;
 
-import static com.meoguri.linkocean.domain.reaction.entity.Reaction.*;
+import static com.meoguri.linkocean.domain.bookmark.entity.Reaction.*;
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
-
-import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
 
 class ReactionTest {
