+++ src/test/java/com/meoguri/linkocean/domain/profile/entity/FollowTest.java
@@ -1,11 +1,9 @@
-package com.meoguri.linkocean.domain.follow.entity;
+package com.meoguri.linkocean.domain.profile.entity;
 
 import static com.meoguri.linkocean.domain.util.Fixture.*;
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
-
-import com.meoguri.linkocean.domain.profile.entity.Profile;
 
 class FollowTest {
